+++ src/lib.rs
@@ -18,6 +18,7 @@ mod slots;
 pub use execution::{CancellationToken, ExecutionOptions};
 pub use gc::{CollectionEvent, CollectionReason, GcStatistics, ManagedHeap};
 pub use slots::SlotReference;
+mod char_categories;
 mod file_io;
 pub mod library;
 pub mod memory;